Finishing refactor of run and compare scripts.
