+++ FNG/create_input.py
