+++ demo_collimator/dagopenmc_run/materials.xml
@@ -1,0 +1,26 @@
+<?xml version='1.0' encoding='utf-8'?>
+<materials>
+  <material id="1" name="Air">
+    <density units="g/cc" value="0.001205" />
+    <nuclide ao="0.018145" name="C12" />
+    <nuclide ao="0.00019625" name="C13" />
+    <nuclide ao="0.34869" name="N15" />
+    <nuclide ao="0.0097798" name="O17" />
+    <nuclide ao="0.052759" name="O18" />
+    <nuclide ao="0.001903" name="Ar36" />
+    <nuclide ao="0.0003588" name="Ar38" />
+    <nuclide ao="0.56817" name="Ar40" />
+  </material>
+  <material id="2" name="Lead">
+    <density units="g/cc" value="11.35" />
+    <nuclide ao="0.014" name="Pb204" />
+    <nuclide ao="0.241" name="Pb206" />
+    <nuclide ao="0.221" name="Pb207" />
+    <nuclide ao="0.524" name="Pb208" />
+  </material>
+  <material id="3" name="Water">
+    <density units="g/cc" value="1.0" />
+    <nuclide ao="2.0" name="H1" />
+    <nuclide ao="1.0" name="O16" />
+  </material>
+</materials>
